Fix of FixeCompg::createUnitVect(Vec3D)

+++ Model/lensModel/makeLens.cxx
@@ -3,7 +3,7 @@
  
  * File:   lensModel/makeLens.cxx
  *
- * Copyright (c) 2004-2015 by Stuart Ansell
+ * Copyright (c) 2004-2016 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -192,7 +192,7 @@ makeLens::createTally(Simulation& System,
   const size_t NSTally=IParam.setCnt("surfTally");
   const size_t NEng=IParam.setCnt("tallyEnergy");
   
-  int FL;
+  size_t FL;
   double Dist;
   for(size_t i=0;i<NSTally;i++)
     {
